+++ src/main/groovy/org/codenarc/rule/unnecessary/UnnecessarySetterRule.groovy
@@ -29,6 +29,7 @@ import org.codenarc.util.AstUtil
  * Setters take one method argument. Setter calls within an expression are ignored.
  */
 class UnnecessarySetterRule extends AbstractAstVisitorRule {
+
     String name = 'UnnecessarySetter'
     int priority = 3
     Class astVisitorClass = UnnecessarySetterAstVisitor
@@ -62,15 +63,10 @@ class UnnecessarySetterAstVisitor extends AbstractAstVisitor {
                 && name[0..2] == 'set'
                 && (name[3] as Character).isUpperCase()
                 && (name.length() == 4 || name[4..-1] != name[4..-1].toUpperCase()) ) {
-            // TODO Restore once CodeNarc upgrades to Groovy 2.4
-            // def propertyName = name[3..-1].uncapitalize()
-            // def assignment = AstUtil.getNodeText(call.arguments, sourceCode)
+            def propertyName = name[3..-1].uncapitalize()
+            def assignment = AstUtil.getNodeText(call.arguments, sourceCode)
 
-            def propertyName = name[3].toLowerCase()
-            if (name.length() > 4) {
-                propertyName += name[4..-1]
-            }
-            addUnnecessarySetterViolation(call, propertyName)
+            addUnnecessarySetterViolation(call, propertyName, assignment)
         }
     }
 
@@ -79,15 +75,12 @@ class UnnecessarySetterAstVisitor extends AbstractAstVisitor {
         objectExpression instanceof VariableExpression && objectExpression.superExpression
     }
 
-    private void addUnnecessarySetterViolation(MethodCallExpression call, String propertyName) {
+    private void addUnnecessarySetterViolation(MethodCallExpression call, String propertyName, String assignment) {
         // Only add if there is not already a field with that name
         def fieldNames = currentClassNode.fields.name
         if (!fieldNames.contains(propertyName)) {
             String name = call.method.value
-            addViolation call, "$name(..) can probably be rewritten as $propertyName = .."
-
-            // TODO Restore once CodeNarc upgrades to Groovy 2.4
-            //addViolation call, "$name($assignment) can probably be rewritten as $propertyName = $assignment"
+            addViolation call, "$name($assignment) can probably be rewritten as $propertyName = $assignment"
         }
     }
 }